WARNING, IN THIS COMMIT, THE PROGRAM IS NOT STABLE!
- Updated Rename PNG Views tool project.
- Added new player shoot assets:
    - Added sprite assets.
    - Create PLYSHOOT.FPG.
- DIV2GJAM.PRG:
    - Major refactorization applied to code.
    - Implemented new player shoot sprites. Setup now as double shot.
    - Implemented typeof() function to get the type of a process instance.
    - Implemented flash() function for ease create flash effects with fade() calls (for any reason, in this program not working properly!).
    - Separated read_input() code in read_keyboard_input() and read_joystick_input() functions.
    - Implemented hyperspace level entry animation with some HUD animation (text only) in screen.
    - MAJOR ISSUE DETECTED!!!
MAYBE WE REACH THE LIMIT OF PROCESSES CREATION IN MEMORY (256 MAX LIMIT IS THE VALUE WE THINK FOUND IN DIV GAME STUDIO 2 SOURCE CODE).
IN THE CURRENT PROGRAM STATE, IF USING DEFAULT VALUES FOR STARFIELD STAR INSTANCES AND BACKGROUND ASTEROIDS, THE PROGRAM CRASHED WITHOUT WARNING.
IN THE BETTER CASES, USING LOWER VALUES, THE PROGRAM CRASHED, FOR EXAMPLE, WHEN THE PLAYER SHOOT FAST MULTIPLE TIMES (WITH THE STARFIELD AND BACKGROUND ASTEROIDS ENABLED).
DESCARTED OUT OF MEMORY ISSUE, memory_free() VALUES LOGGED SHOWS HAS MEMORY FREE ENOUGH FOR THE PROGRAM.
MY TEORY IS IF I REMOVE ALL m7_xgraph() LOGIC, THAT DUPLICATES EVERY PROCESS THAT USED IT, MAYBE REDUCED THE OVERFLOW PROCESS CREATION (NOT SURE IF FUNCTION BLOCKS AFFTED BY THIS ISSUE, MAYBE TOO).
IM AFFRAID TO NEED TO REPROGRAM THE ENTIRE CORE OF THE GAME BECAUSE OF THIS ISSUE.